+++ core/admin/Plugin.php
@@ -5,7 +5,7 @@ namespace Dev4Press\Plugin\CoreActivity\Admin;
 use Dev4Press\Plugin\CoreActivity\Basic\Plugin as CorePlugin;
 use Dev4Press\Plugin\CoreActivity\Basic\Settings as CoreSettings;
 use Dev4Press\Plugin\CoreActivity\Basic\Wizard;
-use Dev4Press\v45\Core\Admin\Network\Plugin as BasePlugin;
+use Dev4Press\v46\Core\Admin\Network\Plugin as BasePlugin;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;