+++ core/admin/panel/Wizard.php
@@ -2,7 +2,7 @@
 
 namespace Dev4Press\Plugin\CoreActivity\Admin\Panel;
 
-use Dev4Press\v45\Core\UI\Admin\PanelWizard;
+use Dev4Press\v46\Core\UI\Admin\PanelWizard;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;