+++ core/basic/Information.php
@@ -2,7 +2,7 @@
 
 namespace Dev4Press\Plugin\CoreActivity\Basic;
 
-use Dev4Press\v45\Core\Plugins\Information as BaseInformation;
+use Dev4Press\v46\Core\Plugins\Information as BaseInformation;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;
@@ -11,11 +11,11 @@ if ( ! defined( 'ABSPATH' ) ) {
 class Information extends BaseInformation {
 	public $code = 'coreactivity';
 
-	public $version = '1.5.4';
-	public $build = 1540;
+	public $version = '1.6';
+	public $build = 1550;
 	public $edition = 'free';
 	public $status = 'stable';
-	public $updated = '2023.12.19';
+	public $updated = '2024.01.03';
 	public $released = '2023.09.06';
 
 	public $github_url = 'https://github.com/dev4press/coreactivity';