+++ core/log/Activity.php
@@ -30,8 +30,8 @@ use Dev4Press\Plugin\CoreActivity\Plugins\Jetpack;
 use Dev4Press\Plugin\CoreActivity\Plugins\SweepPress;
 use Dev4Press\Plugin\CoreActivity\Plugins\UserSwitching;
 use Dev4Press\Plugin\CoreActivity\Plugins\WooCommerce;
-use Dev4Press\v45\Core\Quick\Sanitize;
-use Dev4Press\v45\Core\Quick\Str;
+use Dev4Press\v46\Core\Quick\Sanitize;
+use Dev4Press\v46\Core\Quick\Str;
 use stdClass;
 
 if ( ! defined( 'ABSPATH' ) ) {