+++ core/log/Core.php
@@ -3,13 +3,13 @@
 namespace Dev4Press\Plugin\CoreActivity\Log;
 
 use Dev4Press\Plugin\CoreActivity\Basic\DB;
-use Dev4Press\v45\Core\DateTime;
-use Dev4Press\v45\Core\Helpers\IP;
-use Dev4Press\v45\Core\Quick\Sanitize;
-use Dev4Press\v45\Core\Quick\URL;
-use Dev4Press\v45\Core\Scope;
-use Dev4Press\v45\Service\GEOIP\Location;
-use Dev4Press\v45\WordPress;
+use Dev4Press\v46\Core\DateTime;
+use Dev4Press\v46\Core\Helpers\IP;
+use Dev4Press\v46\Core\Quick\Sanitize;
+use Dev4Press\v46\Core\Quick\URL;
+use Dev4Press\v46\Core\Scope;
+use Dev4Press\v46\Service\GEOIP\Location;
+use Dev4Press\v46\WordPress;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;