+++ core/log/Display.php
@@ -3,7 +3,7 @@
 namespace Dev4Press\Plugin\CoreActivity\Log;
 
 use BP_Groups_Group;
-use Dev4Press\v45\Core\Mailer\Detection;
+use Dev4Press\v46\Core\Mailer\Detection;
 use GFAPI;
 use stdClass;
 use WP_Comment;
@@ -275,6 +275,8 @@ class Display {
 	private function _display_post( stdClass $item ) : string {
 		$render = '';
 
+		switch_to_blog( $item->blog_id );
+
 		$post = get_post( $item->object_id );
 
 		if ( $post instanceof WP_Post ) {
@@ -284,11 +286,15 @@ class Display {
 			$render .= __( 'MISSING', 'coreactivity' ) . ': <strong>' . $item->object_id . '</strong>';
 		}
 
+		restore_current_blog();
+
 		return $render;
 	}
 
 	private function _display_comment( stdClass $item ) : string {
 		$render = '';
+
+		switch_to_blog( $item->blog_id );
 
 		$comment = get_comment( $item->object_id );
 
@@ -306,11 +312,15 @@ class Display {
 			$render .= __( 'MISSING', 'coreactivity' ) . ': <strong>' . $item->object_id . '</strong>';
 		}
 
+		restore_current_blog();
+
 		return $render;
 	}
 
 	private function _display_attachment( stdClass $item ) : string {
 		$render = '';
+
+		switch_to_blog( $item->blog_id );
 
 		$post = get_post( $item->object_id );
 
@@ -320,6 +330,8 @@ class Display {
 		} else {
 			$render .= __( 'MISSING', 'coreactivity' ) . ': <strong>' . $item->object_id . '</strong>';
 		}
+
+		restore_current_blog();
 
 		return $render;
 	}
@@ -400,6 +412,8 @@ class Display {
 	private function _display_term( stdClass $item ) : string {
 		$render = '';
 
+		switch_to_blog( $item->blog_id );
+
 		$term = get_term( $item->object_id );
 
 		if ( $term instanceof \WP_Term ) {
@@ -408,6 +422,8 @@ class Display {
 		} else {
 			$render .= __( 'MISSING', 'coreactivity' ) . ': <strong>' . $item->object_id . '</strong>';
 		}
+
+		restore_current_blog();
 
 		return $render;
 	}
@@ -473,6 +489,8 @@ class Display {
 	private function _brief_post( stdClass $item ) : string {
 		$render = '';
 
+		switch_to_blog( $item->blog_id );
+
 		$post = get_post( $item->object_id );
 
 		if ( $post instanceof WP_Post ) {
@@ -482,11 +500,15 @@ class Display {
 			$render .= __( 'MISSING', 'coreactivity' ) . ': ' . $item->object_id;
 		}
 
+		restore_current_blog();
+
 		return $render;
 	}
 
 	private function _brief_comment( stdClass $item ) : string {
 		$render = '';
+
+		switch_to_blog( $item->blog_id );
 
 		$comment = get_comment( $item->object_id );
 
@@ -504,11 +526,15 @@ class Display {
 			$render .= __( 'MISSING', 'coreactivity' ) . ': ' . $item->object_id;
 		}
 
+		restore_current_blog();
+
 		return $render;
 	}
 
 	private function _brief_attachment( stdClass $item ) : string {
 		$render = '';
+
+		switch_to_blog( $item->blog_id );
 
 		$post = get_post( $item->object_id );
 
@@ -518,6 +544,8 @@ class Display {
 		} else {
 			$render .= __( 'MISSING', 'coreactivity' ) . ': ' . $item->object_id;
 		}
+
+		restore_current_blog();
 
 		return $render;
 	}
@@ -597,6 +625,8 @@ class Display {
 	private function _brief_term( stdClass $item ) : string {
 		$render = '';
 
+		switch_to_blog( $item->blog_id );
+
 		$term = get_term( $item->object_id );
 
 		if ( $term instanceof \WP_Term ) {
@@ -605,6 +635,8 @@ class Display {
 		} else {
 			$render .= __( 'MISSING', 'coreactivity' ) . ': ' . $item->object_id;
 		}
+
+		restore_current_blog();
 
 		return $render;
 	}