+++ core/plugins/BuddyPress.php
@@ -3,8 +3,8 @@
 namespace Dev4Press\Plugin\CoreActivity\Plugins;
 
 use Dev4Press\Plugin\CoreActivity\Base\Plugin;
-use Dev4Press\v45\Core\Quick\Request;
-use Dev4Press\v45\Core\Quick\Sanitize;
+use Dev4Press\v46\Core\Quick\Request;
+use Dev4Press\v46\Core\Quick\Sanitize;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;