+++ core/plugins/GravityForms.php
@@ -3,7 +3,7 @@
 namespace Dev4Press\Plugin\CoreActivity\Plugins;
 
 use Dev4Press\Plugin\CoreActivity\Base\Plugin;
-use Dev4Press\v45\Core\Quick\Sanitize;
+use Dev4Press\v46\Core\Quick\Sanitize;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;