+++ forms/content-update.php
@@ -1,7 +1,7 @@
 <?php
 
-use Dev4Press\v45\Core\Quick\WPR;
-use function Dev4Press\v45\Functions\panel;
+use Dev4Press\v46\Core\Quick\WPR;
+use function Dev4Press\v46\Functions\panel;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;